+++ tests/test_04_dxf_high_level_structs/test_404_tables.py
@@ -41,7 +41,7 @@ class TestAddLayerTableEntry:
             true_color=ezdxf.rgb2int((0x10, 0x20, 0x30)),
             linetype="DASHED",
             lineweight=18,
-            plot=True
+            plot=True,
         )
         assert layer.dxf.name == "NEW_LAYER"
         assert layer.dxf.color == 2
@@ -57,3 +57,41 @@ class TestAddLayerTableEntry:
     def test_check_invalid_line_weight(self, tables: TablesSection):
         with pytest.raises(ValueError):
             tables.layers.add("INVALID_LINE_WEIGHT", lineweight=300)
+
+
+class TestTextStyleTable:
+    def test_add_new_ttf_font_text_style(self, tables: TablesSection):
+        style = tables.styles.add(
+            "NEW_STYLE", font="Arial.ttf", dxfattribs={"flags": 3}
+        )
+        assert style.dxf.name == "NEW_STYLE"
+        assert style.dxf.font == "Arial.ttf"
+        assert style.dxf.flags == 3
+
+    def test_add_new_shape_file(self, tables: TablesSection):
+        style = tables.styles.add_shx("shapes1.shx")
+        assert style.dxf.name == "", "shape files have no name"
+        assert style.dxf.font == "shapes1.shx"
+        assert style.dxf.flags == 1
+
+        # can not add same shape file twice:
+        with pytest.raises(ezdxf.const.DXFTableEntryError):
+            tables.styles.add_shx("shapes1.shx")
+
+    def test_get_shape_file(self, tables: TablesSection):
+        style = tables.styles.get_shx("shapes2.shx")
+        assert style.dxf.name == "", "shape files have no name"
+        assert style.dxf.font == "shapes2.shx"
+        assert style.dxf.flags == 1
+
+        style2 = tables.styles.get_shx("shapes2.shx")
+        assert style is style2
+
+    def test_find_shape_file(self, tables: TablesSection):
+        tables.styles.add_shx("shapes3.shx")
+        style = tables.styles.find_shx("shapes3.shx")
+        assert style.dxf.font == "shapes3.shx"
+
+    def test_if_shape_file_entry_exist(self, tables: TablesSection):
+        assert tables.styles.find_shx("unknown.shx") is None
+