+++ NEWS.md
@@ -13,6 +13,8 @@ Version 0.16.6b2 - dev
   OBJECTS section
 - NEW: `Auditor` removes standalone ATTRIB entities from layouts and blocks
 - NEW: `Drawing.layers.add()` factory method
+- NEW: `Drawing.styles.add()` factory method
+- NEW: `Drawing.styles.add_shx()` factory method
 - CHANGE: renamed `RenderContext.current_layer` to `RenderContext.current_layer_properties` 
 - CHANGE: renamed `RenderContext.current_block_reference` to `RenderContext.current_block_reference_properties` 
 - CHANGE: extended entity validation for `GROUP`